refactor(views): restructure course views into separate files and optimize queryset handling
